+++ __tests__/services/blacklist.ts
@@ -1,6 +1,6 @@
 
 jest.mock('../../src/services/redis')
-import { extractDestinyPhone, isInBlacklistInMemory, addToBlacklistInMemory, cleanBlackList, isInBlacklistInRedis } from '../../src/services/blacklist'
+import { isInBlacklistInMemory, addToBlacklistInMemory, cleanBlackList, isInBlacklistInRedis } from '../../src/services/blacklist'
 import { redisGet, redisKeys, blacklist } from '../../src/services/redis'
 
 const redisGetMock = redisGet as jest.MockedFunction<typeof redisGet>
@@ -8,44 +8,6 @@ const redisKeysMock = redisKeys as jest.MockedFunction<typeof redisKeys>
 const blacklistMock = blacklist as jest.MockedFunction<typeof blacklist>
 
 describe('service blacklist webhook', () => {
-  test('return y extractDestinyPhone from webhook payload message', async () => {
-    const payload = {
-      entry: [
-        {
-          changes: [
-            {
-              value: {
-                contacts: [{ wa_id: 'y' }],
-              },
-            },
-          ],
-        },
-      ],
-    }
-    expect(extractDestinyPhone(payload)).toBe('y')
-  })
-
-  test('return x extractDestinyPhone from webhook payload status', async () => {
-    const payload = {
-      entry: [
-        {
-          changes: [
-            {
-              value: {
-                statuses: [{ recipient_id: 'x' }]
-              }
-            }
-          ]
-        }
-      ]
-    }
-    expect(extractDestinyPhone(payload)).toBe('x')
-  })
-
-  test('return empty extractDestinyPhone from api payload', async () => {
-    expect(extractDestinyPhone({ to: 'y' })).toBe('y')
-  })
-
   test('return false isInBlacklistInMemory', async () => {
     await cleanBlackList()
     expect(await isInBlacklistInMemory('x', 'y', { to: 'w' })).toBe('')
