+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "unoapi-cloud",
-  "version": "1.19.0",
+  "version": "1.19.1",
   "description": "Unoapi Cloud",
   "exports": "./dist/index.js",
   "types": "./dist/index.d.ts",
