+++ src/services/blacklist.ts
@@ -3,6 +3,7 @@ import { amqpEnqueue } from '../amqp'
 import { UNOAPI_JOB_BLACKLIST_ADD } from '../defaults'
 import { blacklist, redisTtl, redisKeys } from './redis'
 import logger from './logger'
+import { extractDestinyPhone } from './transformer'
 
 const DATA = new NodeCache()
 let searchData = true
@@ -13,33 +14,6 @@ export interface addToBlacklist {
 
 export interface isInBlacklist {
   (from: string, webhookId: string, payload: object): Promise<String>
-}
-
-export const extractDestinyPhone = (payload: object) => {
-  const data = payload as any
-  const number = data?.to || (
-    (
-      data.entry
-      && data.entry[0]
-      && data.entry[0].changes
-      && data.entry[0].changes[0]
-      && data.entry[0].changes[0].value
-    ) && (
-      (
-        data.entry[0].changes[0].value.contacts
-        && data.entry[0].changes[0].value.contacts[0]
-        && data.entry[0].changes[0].value.contacts[0].wa_id?.replace('+', '')
-      ) || (
-        data.entry[0].changes[0].value.statuses
-        && data.entry[0].changes[0].value.statuses[0]
-        && data.entry[0].changes[0].value.statuses[0].recipient_id?.replace('+', '')
-      )
-    )
-  )
-  if (!number) {
-    throw Error(`error on get phone number from ${JSON.stringify(payload)}`)
-  }
-  return number
 }
 
 export const blacklistInMemory = (from: string, webhookId: string, to: string) => {
