+++ src/services/transformer.ts
@@ -310,6 +310,33 @@ export const isValidPhoneNumber = (value: string, nine = false): boolean => {
   return !isInValid
 }
 
+export const extractDestinyPhone = (payload: object) => {
+  const data = payload as any
+  const number = data?.to || (
+    (
+      data.entry
+      && data.entry[0]
+      && data.entry[0].changes
+      && data.entry[0].changes[0]
+      && data.entry[0].changes[0].value
+    ) && (
+      (
+        data.entry[0].changes[0].value.contacts
+        && data.entry[0].changes[0].value.contacts[0]
+        && data.entry[0].changes[0].value.contacts[0].wa_id?.replace('+', '')
+      ) || (
+        data.entry[0].changes[0].value.statuses
+        && data.entry[0].changes[0].value.statuses[0]
+        && data.entry[0].changes[0].value.statuses[0].recipient_id?.replace('+', '')
+      )
+    )
+  )
+  if (!number) {
+    throw Error(`error on get phone number from ${JSON.stringify(payload)}`)
+  }
+  return number
+}
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const jidToPhoneNumber = (value: any, plus = '+', retry = true): string => {
   const number = (value || '').split('@')[0].split(':')[0].replace('+', '')
